requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
Django==6.0.3
django-simple-history==3.11.0
idna==3.11
mercadopago==2.3.0
pillow==12.1.1
python-decouple==3.8
qrcode==8.2
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
